interface prep
