+++ packages/designto-code/index.ts
@@ -2,3 +2,6 @@ export * as flutter from "./flutter";
 export * as proc from "./proc";
 export * as react from "./react";
 export * as token from "@designto/token";
+
+/* main function */
+export { designToCode } from "./universal";
