+++ packages/designto-code/proc/build-proc/index.ts
@@ -1,0 +1,15 @@
+import { repo_assets } from "@design-sdk/core";
+
+export abstract class BuildProcess {
+  readonly id: string;
+  readonly processes: Array<BuildProcess> = [];
+  imageAssetRepository: repo_assets.ImageRepository;
+
+  constructor() {
+    this.id = Date.now().toString();
+  }
+
+  registerProcess(process: BuildProcess) {
+    this.processes.push(process);
+  }
+}
