+++ packages/designto-code/proc/configure/build-config.ts
@@ -1,0 +1,5 @@
+import { TargetCodePlatform } from "./target-build-platform";
+
+export interface Configuration {
+  platforms: TargetCodePlatform[];
+}
