+++ packages/designto-code/proc/configure/index.ts
@@ -1,0 +1,2 @@
+export * from "./target-build-platform";
+export * from "./build-config";
