+++ packages/designto-code/proc/configure/target-build-platform.ts
@@ -1,0 +1,10 @@
+export enum TargetCodePlatform {
+  flutter = "flutter",
+  react = "react",
+  vue = "vue",
+  svelte = "svelte",
+  reactnative = "react-native",
+  html = "html",
+  android = "android",
+  ios = "ios",
+}
