+++ packages/designto-code/proc/input/design-input.ts
@@ -1,0 +1,1 @@
+export interface DesignInput {}
