+++ packages/designto-code/proc/input/index.ts
@@ -1,0 +1,1 @@
+export * from "./design-input";
