+++ packages/designto-code/proc/output/code-output.ts
@@ -1,0 +1,1 @@
+export interface CodeOutput {}
