+++ packages/designto-code/universal/design-to-code.ts
@@ -1,0 +1,5 @@
+import { input, output } from "../proc";
+
+export function designToCode(input: input.DesignInput): output.CodeOutput {
+  return;
+}
