+++ packages/designto-code/universal/index.ts
@@ -1,0 +1,1 @@
+export { designToCode } from "./design-to-code";
